Update Solidity imports and formatting in Trade.sol, IVaultManager.sol, AssetVault.sol, and VaultManager.sol

+++ src/VaultManager/IVaultManager.sol
@@ -2,12 +2,18 @@
 pragma solidity =0.8.25;
 
 interface IVaultManager {
-
-    function createVault(address[] memory _initialAssets, uint256 _performanceFee) external;
+    function createVault(address[] memory _initialAssets, uint8 _performanceFee) external;
 
     function addTrader(address _trader) external;
 
     function removeTrader(address _trader) external;
 
-    function executeTrade(uint256 _vaultIndex, address _asset1, uint256 _amount1, address _asset2, uint256 _amount2) external;
+    function executeTrade(
+        uint256 _vaultIndex,
+        address _asset1,
+        uint256 _amount1,
+        address _asset2,
+        uint256 _amount2
+    )
+        external;
 }
